refactor(core): refactor dynamic jni
