+++ android/sdk/src/main/jni/include/bridge/java2js.h
@@ -1,0 +1,40 @@
+/*
+ *
+ * Tencent is pleased to support the open source community by making
+ * Hippy available.
+ *
+ * Copyright (C) 2019 THL A29 Limited, a Tencent company.
+ * All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+#pragma once
+
+#include <jni.h>
+
+namespace hippy {
+namespace bridge {
+
+void CallFunction(JNIEnv* j_env,
+                  jobject j_obj,
+                  jstring j_action,
+                  jbyteArray j_params,
+                  jint j_offset,
+                  jint j_length,
+                  jlong j_runtime_id,
+                  jobject j_callback);
+
+}  // namespace bridge
+}  // namespace hippy
