+++ android/sdk/src/main/jni/include/jni/jni_env.h
@@ -24,37 +24,35 @@
 
 #include <jni.h>
 
-#include "core/core.h"
+#include <memory>
+#include <mutex>
 
 class JNIEnvironment {
  public:
   struct JemthodID_Wrapper {
-    JemthodID_Wrapper() {
-      call_natives_method_id = nullptr;
-      report_exception_method_id = nullptr;
-      inspector_channel_method_id = nullptr;
-      fetch_resource_method_id = nullptr;
-    }
-
-    jmethodID call_natives_method_id;
-    jmethodID report_exception_method_id;
-    jmethodID inspector_channel_method_id;
-    jmethodID fetch_resource_method_id;
+    jmethodID call_natives_method_id = nullptr;
+    jmethodID report_exception_method_id = nullptr;
+    jmethodID inspector_channel_method_id = nullptr;
+    jmethodID fetch_resource_method_id = nullptr;
   };
 
  public:
+  static std::shared_ptr<JNIEnvironment> GetInstance();
+  static bool ClearJEnvException(JNIEnv* env);
+  static void DestroyInstance();
+
   JNIEnvironment() = default;
   ~JNIEnvironment() = default;
 
+  inline JemthodID_Wrapper GetMethods() { return wrapper_; }
   void init(JavaVM* vm, JNIEnv* env);
+  JNIEnv* AttachCurrentThread();
+  void DetachCurrentThread();
 
-  static bool ClearJEnvException(JNIEnv* env);
-  static JNIEnvironment* GetInstance();
-  static void DestroyInstance();
-  static JNIEnv* AttachCurrentThread();
-  static void DetachCurrentThread();
+ private:
+  static std::shared_ptr<JNIEnvironment> instance_;
+  static std::mutex mutex_;
 
- public:
-  JavaVM* jvm_;
+  JavaVM* j_vm_;
   JemthodID_Wrapper wrapper_;
 };
