+++ android/sdk/src/main/jni/src/jni/scoped_java_ref.cc
@@ -28,9 +28,9 @@
 JavaRef::JavaRef(JNIEnv* env, jobject obj) : obj_(nullptr) {
   // HIPPY_DLOG(hippy::Debug, "JavaRef create");
   if (!env) {
-    env = JNIEnvironment::AttachCurrentThread();
+    env = JNIEnvironment::GetInstance()->AttachCurrentThread();
   } else {
-    HIPPY_DCHECK(env == JNIEnvironment::AttachCurrentThread());
+    HIPPY_DCHECK(env == JNIEnvironment::GetInstance()->AttachCurrentThread());
   }
 
   if (obj) {
@@ -41,6 +41,6 @@ JavaRef::JavaRef(JNIEnv* env, jobject obj) : obj_(nullptr) {
 JavaRef::~JavaRef() {
   // HIPPY_DLOG(hippy::Debug, "~JavaRef release");
   if (obj_) {
-    JNIEnvironment::AttachCurrentThread()->DeleteGlobalRef(obj_);
+      JNIEnvironment::GetInstance()->AttachCurrentThread()->DeleteGlobalRef(obj_);
   }
 }
